.NET Code generator rework handling the full calc
