+++ packages/jsii-dotnet-runtime-test/generate.sh
@@ -3,10 +3,7 @@ set -euo pipefail
 test="./test"
 
 # Generate NuGet packages for jsii-calc and its dependencies.
-# TODO: uncomment this for building jsii-calc
-# jsii-pacmak -vv -t dotnet --recurse -o ${test}/generated ../jsii-calc
-jsii-pacmak -vv -t dotnet --recurse -o ${test}/generated ../jsii-calc-base
-jsii-pacmak -vv -t dotnet --recurse -o ${test}/generated ../jsii-calc-lib
+jsii-pacmak -vv -t dotnet --recurse -o ${test}/generated ../jsii-calc
 
 # Generate Directory.Build.props
 /usr/bin/env node ./Directory.Build.props.t.js > ${test}/Directory.Build.props
